Network wrappers refactoring: move persistent connection management from Kuzzle root component to Network wrappers
